Rename `Grouping` to `Group`.
