--- tests/Grouping/Div/AttributeTest.php
@@ -1,115 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace PHPForge\Html\Tests\Grouping\Div;
-
-use PHPForge\Html\Grouping\Div;
-use PHPForge\Support\Assert;
-use PHPUnit\Framework\TestCase;
-
-/**
- * @psalm-suppress PropertyNotSetInConstructor
- */
-final class AttributeTest extends TestCase
-{
-    public function testAttributes(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div class="value">
-            </div>
-            HTML,
-            Div::widget()->attributes(['class' => 'value'])->render()
-        );
-    }
-
-    public function testClass(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div class="value">
-            </div>
-            HTML,
-            Div::widget()->class('value')->render()
-        );
-    }
-
-    public function testContent(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div>
-            value
-            </div>
-            HTML,
-            Div::widget()->content('value')->render()
-        );
-    }
-
-    public function testDataAttributes(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div data-value="value">
-            </div>
-            HTML,
-            Div::widget()->dataAttributes(['value' => 'value'])->render()
-        );
-    }
-
-    public function testId(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div id="value">
-            </div>
-            HTML,
-            Div::widget()->id('value')->render()
-        );
-    }
-
-    public function testLang(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div lang="value">
-            </div>
-            HTML,
-            Div::widget()->lang('value')->render()
-        );
-    }
-
-    public function testName(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div name="value">
-            </div>
-            HTML,
-            Div::widget()->name('value')->render()
-        );
-    }
-
-    public function testStyle(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div style="value">
-            </div>
-            HTML,
-            Div::widget()->style('value')->render()
-        );
-    }
-
-    public function testTitle(): void
-    {
-        Assert::equalsWithoutLE(
-            <<<HTML
-            <div title="value">
-            </div>
-            HTML,
-            Div::widget()->title('value')->render()
-        );
-    }
-}
